requirements.txt
asgiref==3.3.4
boto3==1.17.108
botocore==1.20.108
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-storages==1.11.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==2.10
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
